requirements.txt
humanfriendly==10.0
nextcord==2.3.2
openai==0.26.0
psutil==5.9.4
python-dotenv==0.21.1
requests==2.28.1
